+++ apps/web/public/static/locales/en/common.json
@@ -3711,6 +3711,5 @@
   "trait": "Trait",
   "voice_id": "Voice ID",
   "use_voice": "Use Voice",
-  "language": "Language",
   "ADD_NEW_STRINGS_ABOVE_THIS_LINE_TO_PREVENT_MERGE_CONFLICTS": "↑↑↑↑↑↑↑↑↑↑↑↑↑ Add your new strings above here ↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑↑"
 }
